requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
python-dotenv>=1.0,<2.0
numpy>=1.26,<3.0
nibabel>=5.2,<6.0
Pillow>=10.0,<12.0
gunicorn>=20.1,<21.0
psycopg2-binary>=2.9,<3.0
whitenoise>=6.0,<7.0
dj_database_url
setuptools>=68.0,<69.0
requests
keras>=3.14,<4.0
scipy>=1.11,<2.0
jax>=0.4,<1.0
jaxlib>=0.4,<1.0
supabase>=2.0,<3.0